Updated most of the XML interface to match NAS2D's interface. Refs issue #14.
Still needs a little work. The documentation is very much lacking and tends to use an older style of Doxygen style commenting. Through most of it I've already changed it over to the same style used by the rest of NAS2D.

I've also started moving most of the documentation commenting to the source files instead of the headers, also to conform with the way the NAS2D handles its documentation. This also results in a much easier to read header interface for quick reference with documentation provided separately as standalone HTML files generated by Doxygen.