Remove early Filesystem initialisation test
makeDirectory is only used during the initialisation of Filesystem on linux,
as such FILESYSTEM_INITIALIZED will always be false during this phase,
thus throwing filesystem_not_initialized() erroneously.

As makeDirectory is only used once in the project and only contains one
single line, it could be inlined if there aren't any plans to use it
elsewhere.