Add simple mouse pointer management
The cursors are simply kept track of by a global std::map in
OGL_Renderer.cpp. When a cursor is added, an Id is passed to addCursor,
this Id can then subsequentely by be used to set the current currsor via
setCursor.

When the first cursor is added, it is automatically set as the current
cursor. Overwriting an Id should happen without memory leaking.