Make build folders depend on source files
This is more accurate than a folder scan when folders do not contain
source files. This would be the case for empty folders, or header only
folders. It also avoids increasingly deep folder nesting if the source
folder is the root of the project, and thus includes the build folder.

This is more efficient than doing two file system scans. It also avoids
a possible discrepancy between two scans, should another process
change the filesystem in between.