Add SDL2 source build folder before sdl2-config
The source build is newer than the system packaged libraries (otherwise,
what's the point). Hence the include folder of the source build should
be added to the search path before the system path returned by
sdl2-config.

For platforms that don't need or do a source build, this folder will be
empty or non-existent, and so will be ignored when searching for header
files.