Add Dockerfile for Ubuntu-18.04 base image
This allows Docker to build an image based on Ubuntu-18.04 with the g++
compiler installed, along with necessary dependencies, such as GLEW,
physfs, SDL2, and SDL2 modules for Image, Mixer, and TTF.

Default command for this image is to run "make" to build the code.

Example build and run commands are given in the header comments of the
Dockerfile.