Add maintainer LABEL to Dockerfile
This is being done largely to silence a linter warning.

The MAINTAINER tag is deprecated. The alternative suggestion is to use a
LABEL tag to specify a maintainer.

On a related note, when "maintainers" are specified in the header
comments of source files, this is known to discourage other people from
making changes. This can be bad at large companies where many people may
need to update or maintain a file, and probably even worse for large
open source projects.

If quality control or review is really needed, that is better addressed
at the version control level through commit hooks and code reviews as
part of the usual pull-request workflow.

Hence, the maintainer field is deliberately being set to something
obviously fake and inclusive.