Change maintainer LABEL to MAINTAINER tag
This is being done to silence a linter warning.

The MAINTAINER tag was deprecated in Docker 1.13. When needed in newer
files, the suggestion is to use a LABEL tab with a "maintainer" key.

I think better practice is to just omit the line entirely. Declaring a
maintainer is known to discourage other people from taking up
maintenance tasks.