Move call to temporarily set write directory
By moving the `PHYSFS_setWriteDir` call into the `if` block, it becomes
more clear it's being used specifically for the directory creation.

The intended write directory might not exist, and so will need to be
created. It can be created with the `PHYSFS_mkdir` function. The
`PHYSFS_mkdir` function requires a write directory to be set before
calling, and creates a folder relative to the currently set write
directory. This is part of PHYSFS's security scheme, where no writes or
filesystem modifications can be made outside of the write directory.