Set `mDirSeparator` for all platforms
Previously the `mDirSeparator` member variable was only being set for
Windows or Mac builds (which was the only place it was used).

Potentially this private member variable could be removed, as it is only
used in the one function where it is set. It could easily be a local
variable, or even just a temporary value.