Remove platform guard
An unsupported check isn't really required, and probably not much
desired. Physfs documentation suggests it supports more than the
platforms checked here, making this check overly limiting. Additionally,
if Physfs doesn't support a platform, it should be up to Physfs to
report that error.

In terms of porting, if a platform is not supported, it's very likely to
result in compile errors. Those natural errors are probably more useful
than a generic unsupported message. Better to just give a porter the
errors they are looking for.