Remove SDL2 source folder variables
This was intended to be used to build the modules before completing the
install of SDL2 core file. This did not work. The configure scripts of
the additional SDL2 modules tried to compile a test program, and would
fail when SDL2 was not installed, even if it had been unpacked and
compiled in the source folder first.

It appears the only proper thing to do, is to actually complete the full
installation steps for the core SDL2 package first, then try to
configure and compile the additional SDL2 modules.