Move Docker rules to top level Makefile
This makes it easier to run `docker` commands from the root project folder.

Additionally, this prevents invalidating Docker images with updates to
the `docker` commands. Previously that would happen because the
`docker/Makefile` file was copied into the Docker images to perform the
source build, and so updates to that file would invalidate the Docker
images at the point of the file copy.