Do source install of Physfs on Ubuntu 16.04
This updates the Docker image for Ubuntu 16.04 to do a source install of
Physfs. Currently it is installing Physfs 2.0.3, which matches the Apt
package version. This is to smooth out any kinks with switching to a
source install, such as catching the cmake build dependency was missing.

This also provides an example of how to install an alternate package
version than the source install scripts are set to use by default.
Simply override the source package variable with a new download URL.