Remove bad check before mount
The `PHYSFS_exists` method takes a platform independent path which it
searches the virtual filesystem for. Here the method is being used on a
platform dependent path, which is likely outside the virtual filesystem,
as it is about to be mounted into the virtual filesystem.

Additionally, the error reporting if a `PHYSFS_mount` call fails should
already be enough to diagnose the problem.