Allow non-default-constructible types with `Utility::get`
For default constructible types, the `Utility::get` method will continue
to function as before. If the objet hasn't yet been constructed, a new
default constructed object will be created.

For non-default-constructible types, the method will now throw a runtime
error if it is called and the underlying object has not yet been created
with `Utility::init`.

The decision as to how `nullptr` values are handled is made at
compile-time, based on the type of the pointed to object that `Utility`
is constructed around. It is not a runtime check.