Replace constexpr-if with `enable_if`
This splits the function into two separate overloads. Which overload is
selected is based on whether or not the type `T` is default
constructible.

This is equivalent to the previous constexpr-if code, but using
primitives available in earlier versions of the C++ standard. The
`enable_if` template was added in C++11. The `enable_if_t` alias was
added in C++14.

We could potentially drop requirements all the way down to C++11 by
replacing `enable_if_t` with the slightly more verbose syntax:
```
typename std::enable_if<...>::type
```
Note the addition of both `typename` and `::type`.

The current TavisCI environment uses an older GCC compiler without
support for constexpr-if. To upgrade the GCC compiler (for every build)
would be prohibitively expensive in terms of CI build runtime.