Remove `Filesystem::userPath`
I'm not sure if we need to support this method. It's not used in OPHD.
It's implementation uses a deprecated PhysFS method. Do we need this?

Closes #84.

I don't mind providing a method such as this, but I think we need some
clarity on what it is we want to support. In particular, do we want a
`uesrPath` method, which might not be so useful for a game, but probably
useful for a graphics editor and picture viewing program. Or do we want
a `writePath` method, which returns the real filesystem path of the
write folder? That might be useful in reporting paths to the end user,
or letting them know where their files are stored. The previous code
conflated these two ideas.