Separate out build-and-test steps into command
This can be re-used between Linux and MacOS.

This gives a single point of update for building and testing on
Unix-like platforms (Linux, MacOS).

The `checkout` step was not included in this command as it would differ
between a branch build and a PR build. This is planned for the near
future, and may coincide with retiring the TravisCI builds.