Install Google Test to MacOS environment
This downloads a source package and builds it. After building, needed
files are copied to a build result folder which is cached. If this
folder exists after a cache restore, future builds will skip the
Google Test download and build step.

Cache folder and final install folder are both parameterized.

Note: Parameters are not quoted to allow for expansion of "~/" to the
user's home folder. If quotes are needed, the parameter value can be
quoted instead.