Remove `virtual` keyword
There doesn't seem to be any special use for this keyword on these
methods. Using `virtual` on the derived methods would be useful for
creating future derived classes from this one, rather than requiring
someone to inherit from the base class and start over from there. For
such a design, we should probably use `virtual` consistently on all
derived methods, or on none of them.