Remove `clamp` in favor of `std::clamp`
Explicit template instantiation needs to be used when not all parameters
are of the same type. This can happen when indexing a string to find a
`const char` data type, while the other literal value parameters `255`
are interpreted as `int`. By being explicit with the template
instantiation, the compiler will promote values to the selected type.