Add explicit `default` of special member functions
This removes the Clang warning.

The base class is abstract, so no copy slicing can occur. As such, all
operations are defaulted, so as not to restrict derived types. The
derived types are in control as to whether copy/move special member
functions should be allowed.

Historical note: Previously there was no `MixerNull` class, so the base
'Mixer` class doubled for that use.