Update Docker image with newer GoogleTest library
This requires doing a source install of GoogleTest. The Apt packaged
library is only up to version 1.8, whereas the current release is 1.10.

The 1.10 release introduced the new `MOCK_METHOD` macro. This is much
easier to use than the older `MOCK_METHOD0` style macros, where the
number of parameters has to be manually counted and specified.