Remove Dockerfile for Ubuntu 16.04 builds
This image is out of date and no longer functions. The code has been
updated to use C++17 features, such as the use of `<string_view>` in the
StringUtils.h file. This feature is not supported by older C++
compilers, which don't support C++17, such as the stock GCC compiler on
Ubuntu 16.04.

The Ubuntu 18.04 LTS build environment is much more mainstream now
anyway, so we should probably stick with that. Plus, the Ubuntu 20.04
LTS release should be coming out in just a few months.

The source install scripts may still have some use locally, particularly
for testing with updated versions of dependencies. It has been verified
to still work locally.