Add Rectangle factory functions
Convenience methods for creating a Rectangle from Points and Vectors.

This is done as a factory function rather than a constructor. Having a
user defined constructor makes a type non-aggreegate, which makes it
non-POD. We would like to keep these types as simple POD types.
Eventually all constructors will be removed.