Add method `Vector::to<NewBaseType>()`
This provides a shorter easier way to convert a `Vector` to a new
base type. Client code no longer needs to explicitly use `static_cast`
and only has to specify the new base type, not the full struct name
specialized for the new base type. It also avoids having to specify the
namespace.