Add `dataPath` mount relative to `basePath()`
The `dataPath` is now mounted both relative to the current working
directory, and relative to the `basePath()`. This allows for flexibility
in packaging and modding. Static assets may be found relative to the
executable for distribution and install purposes, or found relative to
the current working directory for development and modding purposes.

The mount relative to the current working direction takes priority over
the mount relative to the `basePath()`. This allows for easier modding,
as a shared executable can be used with a new `dataPath` folder for
static assets.