Move `-Werror` flag to environment variable
This enables warnings as errors on CI builds, but does not enable it by
default for local builds. This means developers can still work with a
distinction of warnings from errors. This may be a useful diagnostic
distinction, and allows developers to temporarily test local code that
still produces warnings. Such warnings would of course cause CI builds
to fail, but as long as the warnings are fixed before pushing, it won't
matter.