Add `<SDL2/SDL_Image.h>` include to ImageInfo
I assume this was not included, and a `void*` field was used to reduce
name pollution, and bring in less code for a faster compile. However,
the only two uses of ImageInfo.h are from Image.cpp and
RendererOpenGL.cpp, both of which already include SDL_Image. As such,
there are no realized advantages. It would be surprising if there were,
since to use the field, you need to have the proper types visible, and
the struct isn't very useful without using that one field. As such,
using `void*` mostly just leads to needless casting.

Better to add the include and use a properly typed field.