Specify C++ compiler for Physfs build step
Physfs doesn't actually use a C++ compiler, however the build the can
still fail if it tries to find a C++ compiler and can't find a working
one. This hasn't been an issue for the build image, though if the steps
are repeated in a non-docker environment, where CXX may be set to a
different compiler, a failure can be observed. Although the change is
not strictly required for the Docker image to work, it perhaps serves as
useful documentation for anyone trying to replicate the Docker
environment.