Enable warning `-Wextra`
Add a partial exclusion for GCC-8. The GCC-8 compiler turns on a number
of flags for this blanket warning, including one that warns for function
casts used in the Delegate code. As it doesn't look like that code can
or will be fixed anytime soon, we'll disable that warning for now.
Meanwhile, we can benefit from the other additional checks.