Reverse order of handler init/deinit
It seems slightly cleaner to setup event handling before asking the SDL
Mixer subsystem to start delivering events. Similarly we should ask the
SDL Mixer subsystem to stop delivering events before we remove the event
handlers.

Not that it really matters though, since the Signal class is safe to
call when there are no handlers setup to delegate to.