Fix warnings from `-Wcomma`
We should probably avoid multiple statements on one line. I wanted to
keep this minimal though, and I can see some value in the existing
formatting.

Perhaps we can come up with a vectorized way to handle these operations
in the future that could condense the code, while keeping it expressive.