Update SDL2-Mixer's dependency options
Add support for more optional audio formats. In particular, for OGG
Vorbis. To get the install to work, we need to use the dynamic load
option. This requires loading the "vorbisfile.dll" file at runtime to
process OGG files.

The "vorbisfile.dll" might not be copied by Vcpkg dependency management
during build, so that may need to be a manual step. The DLL has
dependencies of it's own, so the following files need to be copied for
an executable that wishes to use OGG Vorbis files:
- `vorbisfile.dll`
- `vorbis.dll`
- `ogg.dll`