Use `std::ptrdiff_t` for File seeking
Technically these offsets are relative (from either the beginning or end
of the file), and so are difference types. A difference type is used to
offset an iterator to get an iterator at a new position.

As an alternative, we could also cast to `std::ptrdiff_t` within the
function. That would make sense for the forward seek, as the relative
position coincides with the absolute position. It's a bit awkward for
the reverse seek though, where the two are different, and the concept of
relative offset is very much apparent.