Fix sign conversion warning in Font string width calculation
I guess technically the `std::clamp` is still required, since the C++
standard allows types to be larger than the minimum required size.

Note that we should convert to an unsigned 8-bit type before the clamp,
since otherwise we risk losing half the codes. A `char` type is not
specified by the standard to be either signed or unsigned. If it is
unsigned, there would be no problem. If it was signed, then codes
-128..-1 could be mapped in unexpected ways. Taken as raw signed values,
-128..-1 could be clamped to 0. Converted to an unsigned type greater
than 8-bits first, values -128..-1 would be mapped to large unsigned
quantities, and would then be clamped to 255. We want all 8-bit codes to
be guaranteed to be in the range 0..255 before they are clamped.

If the code is outside the range -128..127 (unsigned) or 0..255
(signed), then we're in the realm of having no original 8-bit type to
begin with, so whatever codes are outside that range would be clamped in
an undefined way.