Use `std::string::assign` overload taking first and last
This other overload is simpler to use here, and avoids a sign conversion
warning. The result of a pointer difference is `std::ptrdiff_t`, which
is a signed type, while the overload that takes a start address and
count expects a `std::size_t` for the count.

Checked documentation to ensure the alternative overload behaves as
expected.  Manually verified results to ensure they were identical.