Add Vector::lengthSquared() method
This is often used in distance calculations. Sometimes the square is
sufficient when checking if a distance boundary has been crossed. Using
the square is more efficient in those cases, since finding the actual
distance with a square root operation is expensive. It also potentially
introduces floating point, or a custom integer routine, which is
something I wish to avoid for now in a Vector with a generic BaseType.