Remove joinSkipEmpty method
There appears to be no uses of this method. If we have use for the code,
we can always bring it back from history.

Additionally, the method appears to be highly specialized, so may not be
the most appropriate for the library. There is also a sufficient
workaround, in that a vector can be filtered of empty elements first,
before a regular join is performed. This may have slightly lower
efficiency, though is hard to tell without measuring.