Remove splitSkipEmpty
This method isn't used. If we really need it, we can bring it back from
history.

Additionally, the method seems too highly specialized for the library.
If a workaround is needed, a user could potentially use split, and then
filter the results to remove empty entries. That may be less efficient,
though hard to say how much of an impact it would have without measuring
it.