Use `std::string_view` for `trimWhitespace` parameter
We could potentially also use `std::string_view` for the return value.
The lifetime would be at least as long as the parameter, which would
provide enough time to allow the result to be safely captured, even if
the parameter was a temporary. It would however require explicit capture
to a `std::string` if the parameter was a temporary. Failure to do so
would result in a dangling `std::string_view`. As such, it may be safer
to keep the return value as a `std::string`.