Lift addition for delimiter space out of loop
A single addition at the end should save a bit of work over adding 1 for
the delimiter space each time through the loop.

The calculation is currently off by 1 (both previous and current code).
If you have `n` strings, there are only `n - 1` delimiters needed to
separate them. That becomes a little more obvious in this calculation
form, and is now a little easier to correct it.