Split Configuration::save to allow a filePath parameter
This allows Configuration data to be saved to a file other than one it
was loaded from. In particular, if data is parsed from memory, there
will be no filename set, so it would be necessary to specify one.

Eventually, the loaded filename may become an external concern, and
removed as a class member, along with the `save()` overload that takes
no parameter.

Additionally, the call to `save()` from the destructor may be removed.
It's not particularly safe, since saving may raise exceptions, and
exceptions raised during a destructor call can cause the program to
abort. There is no safe way to propagate exceptions from a destructor.