Create packed rect unpacked imageList primary overload of drawImageRect
The standard containers don't allow elements to be references, hence
there is no way to construct an `ImageList` (`std::vector<Image>`)
from unpacked elements without forcing a copy to be made of the `Image`
objects. Hence, there is no real way for an unpacked imageList overload
to efficiently call a packed imageList overload.