Modify fillVertexArray and fillTextureArray to return arrays
This can be done efficiently with compiler optimizations. Effectively, a
hidden parameter is used to pass a pointer to the target array, where
the caller will write the returned data, and so the callee can fill in
that data directly, without needing to copy data. This optimization may
be more effective when initializing a new variable, as assignment can
produce a copy from a temporary to the final location.