Fix possibly bug in destination size expressions
It appears the code was written to expect Rectangles in the `{point,
point}` format, rather than the `{point, size}` format. Additionally, it
doesn't make sense to subtract a point value from a size value, so
existing code doesn't make a whole lot of sense. The code should likely
be using the size values directly, without adjustment for position.