Remove guard caluse in setFrame
None of the other methods are checked like this, so removing it makes
using this method more consistent.

Additionally, using this method with an invalid mCurrentAction should be
a hard error, rather than something which is silently ignored. Such a
case is a bug, which should crash, and be reported to the developer.