Add Sprite constructor taking an initial action
This is important, since without an initial action set, the
mCurrentAction pointer will be nullptr, and so dereferencing it leads to
undefined behaviour.

Previously, calling Sprite methods that accessed mCurrentAction before
calling `play` with a specific action, would either cause an exception
to be raised, or would create a new empty animation. The creation of the
empty animation may have been accidental, and may or may not have been
safe.