Disable SpriteAnimation copy for safety
The actions collection references the imageSheets collection, so we
should avoid copies as that would invalidate the references. In
paritcular, a copy of the SpriteAnimations struct would copy the
imageSheets to a new collection, and then copy the animations to a new
collection, which unfortunately would still reference the old
imageSheets collection from the original SpriteAnimations. Worse, this
would appear to work correctly so long as the original SpriteAnimations
collection stayed in memory, but would suddenly and perhaps mysteriously
break if the original SpriteAnimations was destructed, thereby freeing
memory for the original imageSheets collection, and invalidating any
references to it.

A move of the SpriteAnimations struct itself would not invalidate the
references, since the original data structures are moved to the new
SpriteAnimation object, without moving the associated imageSheet memory.
Effectively references to the memory are moved, without moving the
memory, which has cached references pointing to it.