Remove Image constructor taking only width and height
There are no methods provided to set an image buffer post creation, so
this constructor makes no sense. It creates an Image object that is not
usable.

If we later extend Image to be able to set or modify the image buffer
post creation, then we might want to consider adding this back.  Until
then though, it's better not to have this constructor here.