Remove default for Font pointSize parameter
This eliminates ambiguity with the bitmap Font constructor when the font
spacing parameter is removed.

Downstream projects always explicitly specify a Font pointSize for TTF
fonts, and it makes sense to always do so.

If there are always at least 2 parameters, we don't really need to mark
the constructor as `explicit` anymore.