Update CircleCI build images
Use Ubuntu 20.04 as the base image for upgraded tools. This applies to
the Clang, Mingw-w64, and GCC (latest) versions.

The unmarked Docker image version is still the stock GCC compiler on
Ubuntu 18.04. Using one older compiler, on an older LTS (Long Term
Support) version of Ubuntu will help ensure we don't break something for
a compiler/system that is likely to still be in use.