Remove default copy/move construction
Rule of zero/three/five. Removes default methods that may be unsafe in
typical contexts, and which may be accidentally called.

The Connection object is likely owned by a parent object, with reference
to the parent object in the delegate. That means if the parent object is
copied, the contained Connection object should not also be copied, but
rather reformed with an updated parent object address for the delegate.

By deleting the copy/move constructor/assignment operator, the parent
object can no longer be default copy/move constructed/assigned, which
eliminates potential errors that would break connections, or leave
connections with dangling pointers to the original destructed objects.
Instead, the developer will get an error about the attempted copy/move
and be forced to write appropriate copy/move operations that will reform
the connection for the new object.