Fix viewport bug in RendererOpenGL::minimumSize
The minimum size is not necessarily the new window size. The size would
only change if the window needed to be enlarged to match the new
minimum. If the window was larger, the existing size would be kept
rather than being decreased to the minimum size.

We can fix this be reading back what the actual window size is after a
new minimum has been set.