Add version tag to cache key for MacOS CI build
The version tag can be used to break the cache and forces a rebuild.
Recently there have been updates to dependency package versions. Trying
to do a partial update results in errors. By breaking the cache, we do a
fresh clean install from a blank slate, and avoid problems with the new
updates.

Note that `brew install` always tries to install the most recent version
of packages. There is no package lock, nor any package version
specification. Instead, when partial updates start to be added, we can
bump the version number to force a cache reload with updated versions of
dependencies.