Remove makefile rules to build Google Test
This is now largely handled in the Docker files in the docker folder.
The code there probably does a better job of it. Plus this code was more
relevant on older distributions, such as Ubuntu 18.04, which aren't used
as much anymore.

If we want to bring back this code at some point, it may fit better in
the Makefile provided in the docker folder.