Remove questionable warning flags `-Wswitch-enum` and `-Wswitch-default`
Using `-Wswitch-enum` means there will always be a warning unless an
enum based switch always defines all switch values. This is a bit silly
for keyboard scan code enums and such.

The `-Wall` flag implies `-Wswitch`, which will warn for an enum based
switch that does not cover a particular case. If not all cases need to
be covered, as `default` label will satisfy that check and avoid a
warning. This is quite reasonable.

The `-Wswitch-default` will warn if a switch does not have a default
label. This may actually be counter productive to the `-Wswitch`
warning, since a `default` label will prevent warnings for uncovered
switch values, such as when an enum is extended with a new value. By not
having a `default` label, we could catch such cases, without generating
spurious warnings.

If someone needs to check for switch statements that don't cover all
named labels, they can manually specify the `-Wswitch-default` flag, and
sort out cases that are innocuous, such as keyboard scan code switch
statements.