Exclude Xml folder from clang-format
There is a plan to remove the Xml folder from the project, and instead
use a package management tool to install TinyXML. As such, there is
little reason to put much effort into updating code style in code that
will soon be removed. Instead, ignore the many code style violations
from that code.