Deprecate old `open` and `write` methods in `Filesystem`
Use the new `read` and `write` methods that work directly with a
`std::string` object for the raw data. This avoids using the `File`
class.

Remove unit test for the old `open` method. This is to avoid triggering
a warning for using a deprecated method. It also makes a bit of sense to
remove unit tests for deprecated methods since we're (soon) no longer
making promises about how they work.