Avoid caching desktop resolution
It's possible for the desktop resolution to change while the program is
running, so we should avoid using a cached value here. It would only
make sense to do that if we also setup event listeners to catch a
desktop resolution change, and if there was a noticable performance
improvement to justify this.