Update helper method to isAnyWindowFlagSet
This checks if any combination (bitwise or) of flags is set. This
assumes flags are single bit bitmasks. If a multi-bit combination flag
is used, it may report true if any component bits are set, rather than
all bits for a given flag. Based on SDL documentation, it appears all
flags are single bit.