Add unit tests for parsing "-1" as unsigned values
The `unsigned int` case doesn't currently work the same as the others.
It throws an exception, while the other integer sizes will parse "-1" as
the largest positive value.

Note that previously we were not defining the behavior for parsing
negative values as unsigned integers. Here we are defining what that
should be, partly based on existing behavior, and partly based on common
expectations concerning overflow and wraparound.