Add AnimationSet::Frame::isStopFrame() helper method
Allow 0 to be used as the new sentinel value. Allow the old sentinel
value of -1 to continue to work while we transition to new value.

The field is unsigned, so a value of 0 fits better than -1. Particularly
since neither 0 nor -1 would have a sensible use as a `frameDelay`.
Further, it will simplify the code if `0` is exluded as a possible delay
for frames that are not stop frames.