Add guards around OpenGL calls in Image destructor
This avoids making calls to OpenGL if OpenGL was not previously called.
This works around the possibility of OpenGL not being initialized, or
not having an SDL OpenGL context set.

This means we can now write unit tests involving the Image object
without first creating a window or an OpenGL context. That's something
we might want to avoid as the unit tests will run in a headless
environment on the Continuous Integration server.