Move initialization list init to constructor body
Convert factory function to tuple of components.

We basically need a tempory to hold the result of the function call, and
then split it up into member fields. There's no way to assign and use
such a tempory in the init list. Without the temporary, you'd need to
call the helper function multiple times, once for each field being
assigned. Probably best to just use the constructor body here.

Not using the init list does lead to some double init of member fields.
The impact of initialing an empty `std::map` and then assigning to it
right after should be fairly minimal.