Remove Ubuntu 18.04 build from CircleCI
As we move towards C++20, we need to drop support for older environments
that do not support C++20. In particular, Ubuntu 18.04 was released
2018, April, so pre-dates the C++20 standard.

It should be noted that Ubuntu 18.04 was a LTS (Long Term Support)
release, so it still considered "current" until 2023. Currently though,
there is the Ubunutu 20.04 LTS release, which most people will have
upgraded to, and soon to be released Ubuntu 22.04 LTS release.