Update Filesystem.read unit test to ignore end of line character
The `Filesystem::read` method is a binary read, it does not perform
conversions to the end of line character. However, git may convert line
endings when checking out text files. As such, we don't know exactly
what line ending we'll see when we read the file. We also don't what
line endings the `\n` in the code will expand to, since it's platform
dependent. Maybe best to only check the prefix of the string and ignore
the line ending.