Allow for move semantics for resourceName on new constructor
We don't use move semantics with the one parameter constructor, since
the read from the file would happen after the `std::move` of the
filename to the first parameter. This would likely result in trying to
read from an empty filename. More to the point, it's probably undefined
behaviour. That is possibly also true if the parameter order was
swapped.