Convert forward declares to includes for better usability
Forward declares make sense when working with pointers and references,
which can be used in their pointer form with incomplete types. Forward
declares are also possible for function prototype parameter and return
values. In the case of parameters, the value being passed will have
already been constructed, so presumably the caller already knows this
type, and has already included or defined the type. For function return
values though, the caller might not already know the type without an
additional include. For usability reasons, it makes sense to include
return types from header files that define functions returning those
types. This prevents the calling code from requiring a separate include
to use that particular method.

See the impact this decision had on `AnimationSet` in this commit for an
example.