Remove `const` qualifier from write/update `Filesystem` methods
These qualifiers put undue restrictions on the implementation of this
class. Here they were added where they could be, for this particular
implementation. As we re-write to not use PhysFS, the implementation
will change. In particular, the level of indirection may change, which
matters since C++ does not use transitive const, so removing a level of
indirection may require stripping a `const` qualifier.

From a higher level perspective, a write or update method is not a
"const" operation, so should not be marked as such. Consider if
`Filesystem` was an interface, with an implementation being an in-memory
filesystem. Could that implementation be written while respecting the
`const` qualifiers of all member functions? That may not be possible if
any write/update methods were marked as "const".