Add helper function `findFirstPath`
This searches a given set of search paths for a given path. If the
concatenation of the search path with the given path exists, then return
that concatenated path.

Note: This method does not attempt to lock access to the search paths.
Putting "../" in the given path can allow this search to escape the
folders specified by the search paths. In particular, this behaviour is
different from how PhysFS uses the search path, which does not allow
access to escape from the preconfigured folders. After some discussion
it was decided this was not a feature of PhysFS we cared to maintain.

Of note, in OPHD, the "/" character is restricted from use in the file
save/load dialog, so that would prevent escaping from the preconfigured
search paths.