Add `Filesystem` constructor without `argv_0` argument
This argument is no longer needed and will be removed. For now, we can
use overloaded constructors and constructor delegation to allow callers
to drop passing this argument, without requiring them to do so. That
will open up a transition period where users of the library can be
adjusted without breaking code.