Skip over caching of problem glyphs
Glyphs of size 0 fail to create a character surface when pre-rendering
characters to a bitmap image. We don't actually care about the reason
why the pre-rendering fails. We just want to skip over glyphs that can
be printed. Effectively we will print an empty glyph in their place,
which will likely also have a size of 0.