Mark single parameter `AnimationSet` constructor as `explicit`
This is important to prevent ambiguity if we want to add a `Sprite`
constructor that takes an explicit cache to use, and to remove the
overload that uses an implicit default cache. Without marking this as
explicit, a `std::string` parameter could be used to construct an
`AnimationSet` and call an unexpected `Sprite` constructor overload,
that would not normally have participated when the first constructor
parameter was a `std::string`.